+++ frontend/src/components/articles/MainArticleTop3.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
-import articleContext from './articleContext';
-import Mainart from './mainart';
+import ArticleContext from '../../constants/articles/ArticleContext';
+import Mainart from './MainArtList';
 import { useNavigate } from 'react-router-dom';
 
-const mainArticle = () => {
+const MainArticleTop3 = () => {
   const navigate = useNavigate();
-  const article = articleContext.slice(0, 3);
+  const article = ArticleContext.slice(0, 3);
   return (
     <div>
       <div className="grid max-w-md grid-cols-1 mx-auto mt-12 lg:max-w-full lg:mt-16 lg:grid-cols-3 gap-x-16 gap-y-12">
@@ -17,4 +17,4 @@ const mainArticle = () => {
   );
 };
 
-export default mainArticle;
+export default MainArticleTop3;
