+++ frontend/src/pages/AntArticleList.tsx
@@ -1,8 +1,9 @@
 import React, { useEffect } from 'react';
-import Headers from '@/components/headers/header';
-import Footer from '@/components/footers/footer';
-import Article from '@/components/articles/article';
+import Headers from '@/components/headers/Header';
+import Footer from '@/components/footers/Footer';
+import Article from '@/components/articles/Article';
 import { useLocation } from 'react-router-dom';
+import SubTtitle from '@/components/titles/SubTtitle';
 const AntArticle = () => {
   const location = useLocation();
 
@@ -13,6 +14,9 @@ const AntArticle = () => {
     <div className="flex flex-col h-screen">
       <Headers />
       <div className="flex-1 pt-28 pb-28">
+        <div className="pt-10">
+          <SubTtitle title1={`Article`} title2={`TeamAnt`} title3={`Article`} />
+        </div>
         <Article />
       </div>
       <Footer />
